requirements.txt
pyspark==3.5.0
pyyaml
requests
python-dotenv
pymysql